wip: move function to corelib
